+++ src/quantity.rs
@@ -13,12 +13,12 @@ pub(crate) trait IsQuantity
 
 #[doc = "A quantity of a unit, supports converting from one unit to another." ]
 #[derive(Copy, Clone)]
-pub struct QuantityBase
+pub struct Quantity
 {
     pub(crate) value: f64,
     pub(crate) unit: UnitDefinition
 }
-impl QuantityBase
+impl Quantity
 {
     ///
     /// Create a new instance of `Quantity` with a given `value` and `unit` 
@@ -31,10 +31,19 @@ impl QuantityBase
     ///
     /// Amount of unit stored in quantity
     /// 
-    #[inline]
+    #[inline]    
     pub fn value(&self) -> f64
     {
         self.value
+    }
+
+    ///
+    /// Get mutable reference to the value for this quantity.
+    /// 
+    #[inline]    
+    pub fn value_mut(&mut self) -> &mut f64
+    {
+        &mut self.value
     }
 
     ///
@@ -47,15 +56,23 @@ impl QuantityBase
     }
 
     ///
+    /// Get mutable reference to the definition for this quantity.
+    /// 
+    #[inline]
+    pub fn definition_mut(&mut self) -> &mut UnitDefinition
+    {
+        &mut self.unit
+    }
+    ///
     /// Convert a quantity from one unit to another
     ///
     #[inline]
-    pub fn convert(&self, unit: Units) -> Result<QuantityBase, RuntimeUnitError>
+    pub fn convert(&self, unit: Units) -> Result<Quantity, RuntimeUnitError>
     {
         self.convert_unit(unit.into())
     }    
     #[inline]
-    pub fn convert_unit(&self, unit: UnitDefinition) -> Result<QuantityBase, RuntimeUnitError>
+    pub fn convert_unit(&self, unit: UnitDefinition) -> Result<Quantity, RuntimeUnitError>
     {
         if self.unit == unit
         {
@@ -89,78 +106,78 @@ impl QuantityBase
     }
 
 }
-impl Debug for QuantityBase
+impl Debug for Quantity
 {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result 
     {
         write!(f, "{} {:?}", self.value(), self.definition())
     }
 }
-impl Mul<f64> for QuantityBase
+impl Mul<f64> for Quantity
 {
-    type Output = QuantityBase;
+    type Output = Quantity;
 
     fn mul(self, rhs: f64) -> Self::Output {
         Self{ value: self.value*rhs, unit: self.unit }
     }
 }
 
-impl Div<f64> for QuantityBase
+impl Div<f64> for Quantity
 {
-    type Output = QuantityBase;
+    type Output = Quantity;
 
     fn div(self, rhs: f64) -> Self::Output {
         Self{ value: self.value/rhs, unit: self.unit }
     }
 }
-impl Mul<QuantityBase> for QuantityBase
+impl Mul<Quantity> for Quantity
 {
-    type Output = QuantityBase;
+    type Output = Quantity;
 
-    fn mul(self, rhs: QuantityBase) -> Self::Output {
+    fn mul(self, rhs: Quantity) -> Self::Output {
         Self{ value: self.value*rhs.value, unit: self.unit*rhs.unit }
     }
 }
-impl Div<QuantityBase> for QuantityBase
+impl Div<Quantity> for Quantity
 {
-    type Output = QuantityBase;
+    type Output = Quantity;
 
-    fn div(self, rhs: QuantityBase) -> Self::Output {
+    fn div(self, rhs: Quantity) -> Self::Output {
         Self{ value: self.value/rhs.value, unit: self.unit/rhs.unit }
     }
 }
-impl Add<QuantityBase> for QuantityBase
+impl Add<Quantity> for Quantity
 {
-    type Output=QuantityBase;
+    type Output=Quantity;
 
-    fn add(self, rhs: QuantityBase) -> Self::Output {
+    fn add(self, rhs: Quantity) -> Self::Output {
         Self { value: self.value + rhs.value, unit: self.unit }
     }
 }
-impl Sub<QuantityBase> for QuantityBase
+impl Sub<Quantity> for Quantity
 {
-    type Output=QuantityBase;
+    type Output=Quantity;
 
-    fn sub(self, rhs: QuantityBase) -> Self::Output {
+    fn sub(self, rhs: Quantity) -> Self::Output {
         Self { value: self.value - rhs.value, unit: self.unit }
     }
 }
 
-impl AddAssign for QuantityBase
+impl AddAssign for Quantity
 {
     fn add_assign(&mut self, rhs: Self) {
         self.value += rhs.value;
     }
 }
 
-impl SubAssign for QuantityBase
+impl SubAssign for Quantity
 {
     fn sub_assign(&mut self, rhs: Self) {
         self.value -= rhs.value;
     }
 }
 
-impl MulAssign for QuantityBase
+impl MulAssign for Quantity
 {
     fn mul_assign(&mut self, rhs: Self) {
         self.value *= rhs.value;
@@ -169,7 +186,7 @@ impl MulAssign for QuantityBase
 }
 
 
-impl DivAssign for QuantityBase
+impl DivAssign for Quantity
 {
     fn div_assign(&mut self, rhs: Self) {
         self.value /= rhs.value;
@@ -177,32 +194,32 @@ impl DivAssign for QuantityBase
     }
 }
 
-impl<T:IsQuantity> Mul<T> for QuantityBase
+impl<T:IsQuantity> Mul<T> for Quantity
 {
-    type Output = QuantityBase;
-    fn mul(self, rhs: T) -> QuantityBase {
-        QuantityBase{ value: self.value*rhs.value(), unit: self.definition()*rhs.definition() }
+    type Output = Quantity;
+    fn mul(self, rhs: T) -> Quantity {
+        Quantity{ value: self.value*rhs.value(), unit: self.definition()*rhs.definition() }
     }
 }
-impl<T:IsQuantity> Div<T> for QuantityBase
+impl<T:IsQuantity> Div<T> for Quantity
 {
-    type Output = QuantityBase;
+    type Output = Quantity;
 
-    fn div(self, rhs: T) -> QuantityBase {
-        QuantityBase{ value: self.value/rhs.value(), unit: self.definition()/rhs.definition() }
+    fn div(self, rhs: T) -> Quantity {
+        Quantity{ value: self.value/rhs.value(), unit: self.definition()/rhs.definition() }
     }
 }
 ///
 /// This only compares magnitudes...
 /// 
-impl PartialOrd<QuantityBase> for QuantityBase
+impl PartialOrd<Quantity> for Quantity
 {
-    fn partial_cmp(&self, other: &QuantityBase) -> Option<core::cmp::Ordering> {
+    fn partial_cmp(&self, other: &Quantity) -> Option<core::cmp::Ordering> {
         self.convert_unchecked(other.unit).partial_cmp(&other.value)
     }
 }
 
-impl PartialEq for QuantityBase
+impl PartialEq for Quantity
 {
     fn eq(&self, other: &Self) -> bool {
         self.unit.base == other.unit.base && self.convert_unchecked(other.unit) == other.value